sh: introduce struct clk_div4_table
This patch introduces struct clk_div4_table. The structure
will be used to keep div4 specific data, and is with this
patch replacing the struct clk_div_mult_table pointer arg
used by the sh_clk_div4_register() functions.

Signed-off-by: Magnus Damm <damm@opensource.se>
Signed-off-by: Paul Mundt <lethal@linux-sh.org>

cherry-picked but below are not updated,
because there are some conflict

	arch/sh/kernel/cpu/sh4a/clock-sh7723.c
	arch/sh/kernel/cpu/sh4a/clock-sh7786.c